refactor: SetupStrategy setup 패키지로 이동
