+++ src/main/java/janggi/controller/JanggiController.java
@@ -3,9 +3,9 @@ package janggi.controller;
 import janggi.domain.Position;
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardGenerator;
-import janggi.domain.command.SetupStrategy;
 import janggi.domain.piece.Piece;
 import janggi.domain.setup.ElephantFormation;
+import janggi.domain.setup.SetupStrategy;
 import janggi.domain.team.BlueTeam;
 import janggi.domain.team.RedTeam;
 import janggi.domain.team.Team;
